LayoutTests/http/tests/security/contentSecurityPolicy/1.1/scriptnonce-scriptsrc-blocked.html

<!DOCTYPE html>
<html>
<head>
<script src='../resources/multiple-iframe-test.js'></script>
<script>
var tests = [
    ['yes', 'script-src 127.0.0.1:8000', 'resources/script.js', 'nonce'],
    ['yes', 'script-src 127.0.0.1:8000; script-nonce nonce', 'resources/script.js', 'nonce'],
    ['no', 'script-src 127.0.0.1:8000; script-nonce nonce', 'resources/script.js', 'notnonce'],
    ['no', 'script-src 127.0.0.1:8000; script-nonce notnonce', 'resources/script.js', 'nonce'],
];
</script>
</head>
<body onload="testExperimentalPolicy()">
  <p>
    None of these scripts should execute even though there are parse errors in the policy.
  </p>
